refactor(io): 优化 `Source.copyTo` 系列函数的默认缓冲区大小赋值
在 `core:io` 模块中简化了缓冲区大小的类型转换逻辑。

- **代码清理**：
    - 移除 `Source.copyTo`、`BufferedSource.copyTo` 及 `Source.copyToSuspend` 函数中 `DEFAULT_BUFFER_SIZE` 冗余的 `.toLong()` 调用。
    - 确保默认参数 `bufferSize` 直接使用已定义的 Long 类型常量，提升代码简洁性。

+++ core/io/src/commonMain/kotlin/ciyin/io/Source.kt
@@ -49,7 +49,7 @@ fun Source.readAll(hashingSink: HashingSink) {
  */
 fun Source.copyTo(
     sink: Sink,
-    bufferSize: Long = DEFAULT_BUFFER_SIZE.toLong(),
+    bufferSize: Long = DEFAULT_BUFFER_SIZE,
     progressInterval: Long = 0,
     onProgress: (Long) -> Boolean = { true }
 ): Long {
@@ -95,7 +95,7 @@ fun Source.copyTo(
  */
 fun BufferedSource.copyTo(
     sink: BufferedSink,
-    bufferSize: Long = DEFAULT_BUFFER_SIZE.toLong(),
+    bufferSize: Long = DEFAULT_BUFFER_SIZE,
     progressInterval: Long = 0,
     onProgress: (Long) -> Boolean = { true }
 ): Long {
@@ -141,7 +141,7 @@ fun BufferedSource.copyTo(
  */
 suspend fun Source.copyToSuspend(
     sink: Sink,
-    bufferSize: Long = DEFAULT_BUFFER_SIZE.toLong(),
+    bufferSize: Long = DEFAULT_BUFFER_SIZE,
     progressInterval: Long = 1024 * 1024,
     onProgress: suspend (Long) -> Unit = {}
 ): Long {
